requirements.txt
Django==4.1
django-environ==0.10.0
psycopg2-binary==2.9.6
djangorestframework==3.14
django-filter==23.3
drf-yasg==1.21.7